requirements.txt
aws-cdk.core
aws-cdk.aws-ec2
aws-cdk.aws-sns
aws-cdk.aws-sns-subscriptions
boto3
aws-cdk.aws-iam
aws-cdk.aws-s3
aws-cdk.aws-s3-deployment
